define consider_auto_trait_candidate_for_opaque_ty and call it from consider_auto_trait_candidate
