+++ compiler/rustc_next_trait_solver/src/solve/assembly/structural_traits.rs
@@ -5,17 +5,21 @@ use derive_where::derive_where;
 use rustc_type_ir::data_structures::HashMap;
 use rustc_type_ir::inherent::*;
 use rustc_type_ir::lang_items::{SolverProjectionLangItem, SolverTraitLangItem};
-use rustc_type_ir::solve::SizedTraitKind;
 use rustc_type_ir::solve::inspect::ProbeKind;
+use rustc_type_ir::solve::{MaybeInfo, NoSolutionOrRerunNonErased, RerunReason, SizedTraitKind};
 use rustc_type_ir::{
-    self as ty, Binder, FallibleTypeFolder, Interner, Movability, Mutability, Region, TypeFoldable,
-    TypeSuperFoldable, Unnormalized, Upcast as _, elaborate,
+    self as ty, Binder, FallibleTypeFolder, Interner, Movability, Mutability, Region,
+    TraitPredicate, TypeFoldable, TypeSuperFoldable, Unnormalized, Upcast as _, elaborate,
 };
 use rustc_type_ir_macros::{TypeFoldable_Generic, TypeVisitable_Generic};
 use tracing::instrument;
 
+use super::Candidate;
 use crate::delegate::SolverDelegate;
-use crate::solve::{AdtDestructorKind, EvalCtxt, Goal, NoSolution};
+use crate::solve::{
+    AdtDestructorKind, BuiltinImplSource, CandidateSource, Certainty, EvalCtxt, Goal, GoalSource,
+    NoSolution, has_only_region_constraints,
+};
 
 // Calculates the constituent types of a type for `auto trait` purposes.
 #[instrument(level = "trace", skip(ecx), ret)]
@@ -104,15 +108,56 @@ where
                 .map(Unnormalized::skip_norm_wip)
                 .collect(),
         )),
+        // Opaque types are already handled earlier
+        _ => unreachable!(),
+    }
+}
 
-        ty::Alias(ty::IsRigid::Yes, ty::AliasTy { kind: ty::Opaque { def_id }, args, .. }) => {
-            // We can resolve the `impl Trait` to its concrete type,
-            // which enforces a DAG between the functions requiring
-            // the auto trait bounds in question.
-            Ok(ty::Binder::dummy(vec![
-                cx.type_of(def_id.into()).instantiate(cx, args).skip_norm_wip(),
-            ]))
+pub(in crate::solve) fn consider_auto_trait_candidate_for_opaque_ty<D, I>(
+    ecx: &mut EvalCtxt<'_, D>,
+    goal: Goal<I, TraitPredicate<I>>,
+    def_id: I::OpaqueTyId,
+    args: I::GenericArgs,
+) -> Result<Candidate<I>, NoSolutionOrRerunNonErased>
+where
+    D: SolverDelegate<Interner = I>,
+    I: Interner,
+{
+    let cx = ecx.cx();
+    let source = CandidateSource::BuiltinImpl(BuiltinImplSource::Misc);
+    if ecx.opaque_accesses.might_rerun() {
+        return match ecx.opaque_accesses.rerun_always(RerunReason::AutoTraitLeakage) {
+            Err(e) => Err(e.into()),
+        };
+    }
+
+    for item_bound in cx.item_self_bounds(def_id.into()).skip_binder() {
+        if item_bound.as_trait_clause().is_some_and(|b| b.def_id() == goal.predicate.def_id()) {
+            return Err(NoSolution.into());
         }
+    }
+
+    let candidate = ecx.probe_trait_candidate(source).enter(|ecx| {
+        let hidden_ty = cx.type_of(def_id.into()).instantiate(cx, args).skip_norm_wip();
+        ecx.add_goal(
+            GoalSource::ImplWhereBound,
+            goal.with(cx, goal.predicate.with_replaced_self_ty(cx, hidden_ty)),
+        )?;
+        ecx.evaluate_added_goals_and_make_canonical_response(Certainty::Yes)
+    });
+
+    // If this hidden-type proof would constrain non-region inference,
+    // treating it as a hard success can reveal the hidden type to the
+    // caller. A hard failure can do the same when it depends on caller
+    // bounds. For now, we are very conservative with caller bounds.
+    let param_env_may_leak_hidden_ty = !goal.param_env.caller_bounds().is_empty();
+    match candidate {
+        Ok(candidate) if has_only_region_constraints(candidate.result) => Ok(candidate),
+        Ok(_) => ecx.forced_ambiguity(MaybeInfo::AMBIGUOUS),
+        Err(NoSolutionOrRerunNonErased::NoSolution(_)) if param_env_may_leak_hidden_ty => {
+            ecx.forced_ambiguity(MaybeInfo::AMBIGUOUS)
+        }
+        Err(err) => Err(err),
     }
 }
